requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
future==0.18.2
idna==2.10
Jinja2==2.11.2
joblib==1.0.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.3
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==6.2.6
mkdocs-material-extensions==1.0.1
nltk==3.5
Pygments==2.7.4
pymdown-extensions==8.1.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
tornado==6.1
tqdm==4.56.0
urllib3==1.26.3
